+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/memo/service/MemoService.kt
@@ -1,0 +1,6 @@
+package com.wafflestudio.toyproject.memoWithTags.memo.service
+
+import org.springframework.stereotype.Service
+
+@Service
+class MemoService
