+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/tag/persistence/TagEntity.kt
@@ -1,0 +1,3 @@
+package com.wafflestudio.toyproject.memoWithTags.tag.persistence
+
+class TagEntity
