+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/tag/service/TagService.kt
@@ -1,0 +1,6 @@
+package com.wafflestudio.toyproject.memoWithTags.tag.service
+
+import org.springframework.stereotype.Service
+
+@Service
+class TagService
