+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/user/contoller/UserController.kt
@@ -1,44 +1,47 @@
-package com.wafflestudio.toyproject.memo_with_tags.user.contoller
+package com.wafflestudio.toyproject.memoWithTags.user.controller
+
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RequestBody
+import org.springframework.web.bind.annotation.RestController
 
 @RestController
-class UserController(
+class UserController {
 
-) {
     @GetMapping("/api/v1/auth/register")
     fun register(@RequestBody request: RegisterRequest): RegisterResponse {
-
+        return RegisterResponse("", "", 0L) // 기본값으로 RegisterResponse 반환
     }
 
     @PostMapping("/api/v1/auth/login")
     fun login(@RequestBody request: LoginRequest): LoginResponse {
-
+        return LoginResponse("", "", 0L) // 기본값으로 LoginResponse 반환
     }
 
     @PostMapping("/api/v1/auth/logout")
     fun logout(@RequestBody request: LogoutRequest) {
-
+        // 반환 타입이 Unit이므로 아무 작업 없이 비워 둬도 OK
     }
 
     @PostMapping("/api/v1/auth/verify-email")
     fun verifyEmail(@RequestBody request: VerifyEmailRequest) {
-
+        // 반환 타입이 Unit이므로 아무 작업 없이 비워 둬도 OK
     }
 
     @PostMapping("/api/v1/auth/forgot-password")
     fun forgotPassword(@RequestBody request: ForgotPasswordRequest) {
-
+        // 반환 타입이 Unit이므로 아무 작업 없이 비워 둬도 OK
     }
 
     @PostMapping("/api/v1/auth/reset-password")
     fun resetPassword(@RequestBody request: ResetPasswordRequest) {
-
+        // 반환 타입이 Unit이므로 아무 작업 없이 비워 둬도 OK
     }
 
     @PostMapping("/api/v1/auth/refresh-token")
     fun refreshToken(): RefreshTokenResponse {
-
+        return RefreshTokenResponse("", "", 0L) // 기본값으로 RefreshTokenResponse 반환
     }
-
 }
 
 data class RegisterRequest(
