--- src/main/kotlin/com/wafflestudio/toyproject/memo_with_tags/memo/service/MemoService.kt
@@ -1,7 +1,0 @@
-package com.wafflestudio.toyproject.memo_with_tags.memo.service
-
-import org.springframework.stereotype.Service
-
-@Service
-class MemoService {
-}
