--- src/main/kotlin/com/wafflestudio/toyproject/memo_with_tags/tag/persistence/TagEntity.kt
@@ -1,4 +1,0 @@
-package com.wafflestudio.toyproject.memo_with_tags.tag.persistence
-
-class TagEntity {
-}
