--- src/main/kotlin/com/wafflestudio/toyproject/memo_with_tags/user/service/UserService.kt
@@ -1,4 +1,0 @@
-package com.wafflestudio.toyproject.memo_with_tags.user.service
-
-class UserService {
-}
