+++ src/test/kotlin/com/wafflestudio/toyproject/memoWithTags/MemoWithTagsApplicationTests.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.toyproject.memo_with_tags
+package com.wafflestudio.toyproject.memoWithTags
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
@@ -6,8 +6,7 @@ import org.springframework.boot.test.context.SpringBootTest
 @SpringBootTest
 class MemoWithTagsApplicationTests {
 
-	@Test
-	fun contextLoads() {
-	}
-
+    @Test
+    fun contextLoads() {
+    }
 }
